Update to gulp 4.x export syntax
