+++ lib/deleteIfExists.js
@@ -1,0 +1,7 @@
+const fs = require('fs');
+
+module.exports = function deleteIfExists(path) {
+  if (fs.existsSync(path)) {
+    fs.unlinkSync(path);
+  }
+};
