--- lib/flags.js
@@ -1,37 +1,0 @@
-const fs = require('fs');
-const paths = require('./paths');
-const utils = require('./utils');
-
-/**
- * Writes out the flag file generated by concentrate
- *
- * This is needed so SiteHtmlHeadEntrySetDisplayer will use the compiled
- * LESS paths when rendering the page on the server.
- */
-function writeFlags(error, complete) {
-  const ts = Math.floor(new Date().getTime() / 1000);
-  fs.writeFileSync(paths.compiledFlag, ts);
-  if (complete) {
-    complete();
-  }
-}
-
-module.exports = {
-  /**
-   * Removes the flag file generated by concentrate
-   *
-   * This is needed so SiteHtmlHeadEntrySetDisplayer will switch back to
-   * dynamic LESS compilation
-   */
-  remove: function removeFlags(error, complete) {
-    utils.deleteIfExists(paths.compiledFlag);
-    utils.deleteIfExists(paths.minifiedFlag);
-    utils.deleteIfExists(paths.combinedFlag);
-    if (complete) {
-      complete();
-    }
-  },
-  task: function writeFlagsTask() {
-    return writeFlags();
-  },
-};
