+++ lib/getSymlinks.js
@@ -1,0 +1,11 @@
+const minimist = require('minimist');
+
+module.exports = function getSymlinks() {
+  const knownOptions = {
+    string: 'symlinks',
+    default: { symlinks: '' },
+  };
+
+  const options = minimist(process.argv.slice(2), knownOptions);
+  return options.symlinks;
+};
