+++ lib/paths.js
@@ -33,4 +33,5 @@ module.exports = {
   compiledFlag: 'www/.concentrate-compiled',
   minifiedFlag: 'www/.concentrate-minified',
   combinedFlag: 'www/.concentrate-combined',
+  symlinkSuffix: '.original',
 };
