+++ lib/phplintStream.js
@@ -3,7 +3,6 @@ const log = require('fancy-log');
 const colors = require('ansi-colors');
 const gulpSpawn = require('gulp-spawn');
 const Stream = require('stream');
-const paths = require('./paths');
 
 function spawn() {
   return gulpSpawn({
@@ -50,7 +49,7 @@ function filter(error, err) {
   return filteredStream;
 }
 
-function stream(streamedPaths) {
+module.exports = function phplintStream(streamedPaths) {
   return gulp
     .src(streamedPaths, { cwdbase: true, buffer: true })
     .pipe(spawn())
@@ -59,11 +58,4 @@ function stream(streamedPaths) {
         log(colors.red('[PHP]'), filename, '->', contents);
       })
     );
-}
-
-module.exports = {
-  task: function phplintTask() {
-    return stream(paths.php);
-  },
-  stream,
 };
