--- lib/utils.js
@@ -1,9 +1,0 @@
-const fs = require('fs');
-
-module.exports = {
-  deleteIfExists: function deleteIfExists(path) {
-    if (fs.existsSync(path)) {
-      fs.unlinkSync(path);
-    }
-  },
-};
