+++ tasks/removeFlags.js
@@ -1,0 +1,17 @@
+const paths = require('../lib/paths');
+const deleteIfExists = require('../lib/deleteIfExists');
+
+/**
+ * Removes the flag file generated by concentrate
+ *
+ * This is needed so SiteHtmlHeadEntrySetDisplayer will switch back to
+ * dynamic LESS compilation
+ */
+module.exports = async function removeFlags(error, complete) {
+  deleteIfExists(paths.compiledFlag);
+  deleteIfExists(paths.minifiedFlag);
+  deleteIfExists(paths.combinedFlag);
+  if (complete) {
+    complete();
+  }
+};
