+++ tasks/writeFlags.js
@@ -1,0 +1,16 @@
+const fs = require('fs');
+const paths = require('../lib/paths');
+
+/**
+ * Writes out the flag files generated by concentrate
+ *
+ * This is needed so SiteHtmlHeadEntrySetDisplayer will use the compiled
+ * LESS paths when rendering the page on the server.
+ */
+module.exports = async function writeFlags(error, complete) {
+  const ts = Math.floor(new Date().getTime() / 1000);
+  fs.writeFileSync(paths.compiledFlag, ts);
+  if (complete) {
+    complete();
+  }
+};
